+++ scripts/schema.sql
@@ -3,7 +3,7 @@ CREATE TABLE
     id serial NOT NULL,
     title character varying(255) NOT NULL,
     description text NULL,
-    rating smallint NOT NULL DEFAULT 0,
+    rating NUMERIC(3, 1) NOT NULL DEFAULT 0,
     image character varying(255) NULL,
     created_at timestamp without time zone NOT NULL DEFAULT now(),
     updated_at timestamp without time zone NOT NULL DEFAULT now(),
